Test meta parser 2.2.2.1145 new

New test cases
TestMetaParser/TestCases/list_file_blocks_Case.py
TestMetaParser/TestCases/get_vol_usage_Case.py
TestMetaParser/TestCases/behaviorCase.py

TestMetaParser/TestCases/Utils/metaParserAdapter.py
1. Add 2 new API adapter
2. Add logger to record function result to file

docker config in TestMetaParser/docker
1. functional_test/start_test_meta_parser.sh to start test in docker
2. run TestMetaParser/start_test_in_docker.sh in docker to install meta parser and start test

TestMetaParser/test_data_generator -> generate test data in TestMetaParser/TestCases/test_data_v2
Need connection to phone device.(Change device id in testMetaDataGenerator.py main)
Generation steps
1. Force management app to get new token (send token expired event via socket to event notify server in management app) (socket directory)
2. Extract logcat to get new token(LogcatUtils.py), extract container from hcfs.conf(HCFSConf.py)
3. Use token to download test data from swift via curl (swift.py)
4. Stat expected test results via adb

See merge request !533

+++ tests/functional_test/start_test_meta_parser.sh
@@ -1,15 +1,23 @@
 #!/bin/bash
+# keyword=test-max-pin
 
+echo -e "\n======== ${BASH_SOURCE[0]} ========"
 repo="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && while [ ! -d .git ] ; do cd ..; done; pwd )"
+here="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source $repo/utils/common_header.bash
+cd $repo
 
+echo "########## Setup Test Env"
 $repo/utils/setup_dev_env.sh -m docker_host
+. $repo/utils/env_config.sh
 
-rm -r $repo/dist
-
+echo "########## pi_tester.py"
+if [ -d $repo/dist ]; then
+	rm -r $repo/dist
+fi
 $repo/build.sh pyhcfs
 
+echo "########## pi_tester.py"
 cd $repo/tests/functional_test/TestCases/TestMetaParser/docker
-
-# python prepare.py
-
+umask 000
 python startMetaParserDockerTest.py
